+++ bootstrap_server/main.go
@@ -11,133 +11,16 @@ package main
 import (
 	"context"
 	"flag"
-	"fmt"
 	"log"
-	"time"
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/x509"
-	"encoding/pem"
-	"io/ioutil"
-	"strings"
-	"os"
-	pb "orcanet/market"
-	"github.com/golang/protobuf/proto"
 	"github.com/libp2p/go-libp2p"
 	dht "github.com/libp2p/go-libp2p-kad-dht"
 	record "github.com/libp2p/go-libp2p-record"
-	"github.com/libp2p/go-libp2p/core/crypto"
 	"github.com/libp2p/go-libp2p/core/host"
 	"github.com/libp2p/go-libp2p/core/peer"
-	drouting "github.com/libp2p/go-libp2p/p2p/discovery/routing"
-	dutil "github.com/libp2p/go-libp2p/p2p/discovery/util"
 	"github.com/multiformats/go-multiaddr"
-	"regexp"
-	"errors"
+	"orcanet/util"
+	"orcanet/validator"
 )
-
-//gRPC server 
-type server struct {
-	pb.UnimplementedMarketServer
-}
-
-// Create a record validator to store our own values within our defined protocol
-type OrcaValidator struct{}
-
-//Validates keys and values that are being put into the DHT.
-//Keys must conform to a sha256 hash
-//Values must conform the specification in /server/README.md
-func (v OrcaValidator) Validate(key string, value []byte) error{
-	// verify key is a sha256 hash
-    hexPattern := "^[a-fA-F0-9]{64}$"
-    regex := regexp.MustCompile(hexPattern)
-    if !regex.MatchString(strings.Replace(key, "orcanet/market/", "", -1)) {
-		return errors.New("Provided key is not in the form of a SHA-256 digest!")
-	}
-
-	pubKeySet := make(map[string] bool)
-
-	for i := 0; i < len(value) - 4; i++ {
-		messageLength := uint16(value[i + 1]) << 8 | uint16(value[i])
-		digitalSignatureLength := uint16(value[i + 3]) << 8 | uint16(value[i + 2])
-		contentLength := messageLength + digitalSignatureLength
-		user := &pb.User{}
-
-		err := proto.Unmarshal(value[i + 4:i + 4 + int(messageLength)], user) //will parse bytes only until user struct is filled out
-		if err != nil {
-			return err
-		}
-
-		if pubKeySet[string(user.GetId())] == true {
-			return errors.New("Duplicate record for the same public key found!")
-		}else{
-			pubKeySet[string(user.GetId())] = true
-		}
-
-		userMessageBytes := value[i + 4:i + 4 + int(messageLength)]
-
-		publicKey, err := crypto.UnmarshalRsaPublicKey(user.GetId())
-		if err != nil{
-			return err
-		}
-
-		signatureBytes := value[i + 4 + int(messageLength):i + 4 + int(contentLength)]
-		valid, err := publicKey.Verify(userMessageBytes, signatureBytes) //this function will automatically compute hash of data to compare signauture
-		
-		if err != nil {
-			return err
-		}
-
-		if !valid {
-			return errors.New("Signature invalid!")
-		}
-
-		i = i + 4 + int(contentLength) - 1
-	}
-
-
-    currentTime := time.Now().UTC()
-    unixTimestamp := currentTime.Unix()
-    unixTimestampInt64 := uint64(unixTimestamp)
-
-	suppliedTime := uint64(0)
-	for i := 1; i < 5; i++ {
-		suppliedTime = suppliedTime | uint64(value[len(value) - i]) << (i - 1)
-	}
-
-	if(suppliedTime > unixTimestampInt64){
-		return errors.New("Supplied time cannot be less than current time")
-	}
-
-	return nil
-}
-
-//We will select the best value based on the longest, latest, valid chain
-func (v OrcaValidator) Select(key string, value [][]byte) (int, error){
-	max := 0
-	maxIndex := 0
-	latestTime := uint64(0);
-	for i := 0; i < len(value); i++ {
-		if len(value[i]) > max {
-			//validate chain 
-			if v.Validate(key, value[i]) == nil {
-				
-				suppliedTime := uint64(0)
-				for j := 1; j < 5; j++ {
-					suppliedTime = suppliedTime | uint64(value[i][len(value[i]) - j]) << (j - 1)
-				}
-
-				if(suppliedTime > latestTime){
-					max = len(value[i])
-					latestTime = suppliedTime;
-					maxIndex = i;
-				}
-			}
-		}
-	}
-
-	return maxIndex, nil;
-}
 
 func main() {
 	var bootstrapPeer string
@@ -146,7 +29,7 @@ func main() {
 
 	ctx := context.Background()
 
-	privKey, err := checkOrCreatePrivateKey("privateKey.pem");
+	privKey, err := util.CheckOrCreatePrivateKey("privateKey.pem");
 	if(err != nil){
 		panic(err);
 	}
@@ -172,7 +55,7 @@ func main() {
 	// client because we want each peer to maintain its own local copy of the
 	// DHT, so that the bootstrapping node of the DHT can go down without
 	// inhibiting future peer discovery.
-	var validator record.Validator = OrcaValidator{}
+	var validator record.Validator = validator.OrcaValidator{}
 	var options []dht.Option
 	options = append(options, dht.Mode(dht.ModeServer))
 	options = append(options, dht.ProtocolPrefix("orcanet/market"), dht.Validator(validator))
@@ -191,11 +74,21 @@ func main() {
 	if bootstrapPeer != "" {
 		connectToBootstrapPeer(bootstrapPeer, host, ctx)
 	} 
-	go discoverPeers(ctx, host, kDHT, "orcanet/market")
+	go util.DiscoverPeers(ctx, host, kDHT, "orcanet/market")
 
 	select {}
 }
 
+/*
+ *
+ * Connect to a bootstrap peer using the specified multiaddr string.
+ *
+ * Parameters:
+ *   multiAddr: Multiaddr string of the peer you are trying to connect to.
+ *   host: libp2p host
+ *   ctx: the context
+ * 
+ */
 func connectToBootstrapPeer(multiAddr string, host host.Host, ctx context.Context){
 	peerAddr, err := multiaddr.NewMultiaddr(multiAddr)
 	if err != nil {
@@ -209,82 +102,4 @@ func connectToBootstrapPeer(multiAddr string, host host.Host, ctx context.Contex
 			log.Println("Connection established with bootstrap node:", *peerinfo)
 		}
 	}()
-}
-
-func discoverPeers(ctx context.Context, h host.Host, kDHT *dht.IpfsDHT, advertise string) {
-	routingDiscovery := drouting.NewRoutingDiscovery(kDHT)
-	dutil.Advertise(ctx, routingDiscovery, advertise)
-
-	// Look for others who have announced and attempt to connect to them
-	for {
-		peerChan, err := routingDiscovery.FindPeers(ctx, advertise)
-		if err != nil {
-			panic(err)
-		}
-		for peer := range peerChan {
-			if peer.ID == h.ID() {
-				continue // No self connection
-			}
-			err := h.Connect(ctx, peer)
-			if err != nil {
-				fmt.Printf("Failed connecting to %s, error: %s\n", peer.ID, err)
-			}
-		}
-		time.Sleep(time.Second * 5)
-	}
-}
-
-func checkOrCreatePrivateKey(path string) (crypto.PrivKey, error) {
-	// Check if the privateKey.pem exists
-	_, err := os.Stat(path)
-	if os.IsNotExist(err) {
-		// No private key file, so let's create one
-		privKey, err := rsa.GenerateKey(rand.Reader, 2048)
-		if err != nil {
-			return nil, err
-		}
-		privKeyBytes := x509.MarshalPKCS1PrivateKey(privKey)
-		privKeyPEM := pem.EncodeToMemory(&pem.Block{
-			Type:  "RSA PRIVATE KEY",
-			Bytes: privKeyBytes,
-		})
-		err = ioutil.WriteFile(path, privKeyPEM, 0600)
-		if err != nil {
-			return nil, err
-		}
-		log.Println("New private key generated and saved to", path)
-
-		libp2pPrivKey, _, err := crypto.KeyPairFromStdKey(privKey);
-		if(err != nil){
-			return nil, err
-		}
-
-		return libp2pPrivKey, nil;
-	} else if err != nil {
-		// Some other error occurred when trying to read the file
-		return nil, err
-	}
-
-	// Private key file exists, let's read it
-	privKeyBytes, err := ioutil.ReadFile(path)
-	if err != nil {
-		return nil, err
-	}
-	block, _ := pem.Decode(privKeyBytes)
-	if block == nil || block.Type != "RSA PRIVATE KEY" {
-		log.Println("Private key file is of invalid format")
-		return nil, errors.New("private key file is of invalid format")
-	}
-	privKey, err := x509.ParsePKCS1PrivateKey(block.Bytes)
-	if err != nil {
-		return nil, err
-	}
-	log.Println("Existing private key loaded from", path)
-
-	libp2pPrivKey, _, err := crypto.KeyPairFromStdKey(privKey);
-	if(err != nil){
-		return nil, err
-	}
-
-	return libp2pPrivKey, nil;
 }
